Amazon/requirements.txt
django==1.8
djangorestframework
drf-extensions
django-dcors